Modifications sur les propriétés des aliens
