+++ Spatial-Invasor/Spatial-Invasor/Octopus.cs
@@ -7,17 +7,10 @@ namespace SpatialInvasor
 {
     public class Octopus : Alien
     {
-        bool Direction = true;
+        
 
         public Octopus(MainGame game) : base(game)
         {
-            Speed = 250f;
-
-            ShootingSpeed = 900f;
-
-            Position = new Vector2(250, 100);
-            Limits = new float[2] { 250f, 700f };
-
             SheetPositions = new List<Rectangle>()
             {
                 new Rectangle(36, 1, 36, 24)
