+++ Spatial-Invasor/Spatial-Invasor/Squid.cs
@@ -7,16 +7,9 @@ namespace SpatialInvasor
 {
     class Squid : Alien
     {
-        bool Direction = true;
+        
         public Squid(MainGame game) : base(game)
         {
-            Speed = 250f;
-
-            ShootingSpeed = 900f;
-
-            Position = new Vector2(250, 200);
-            Limits = new float[2] { 250f, 700f };
-
             SheetPositions = new List<Rectangle>()
             {
                 new Rectangle(73, 1, 24, 24)
