+++ Spatial-Invasor/Spatial-Invasor/bin/Debug/netcoreapp3.1/Spatial-Invasor.runtimeconfig.dev.json
@@ -1,8 +1,9 @@
 {
   "runtimeOptions": {
     "additionalProbingPaths": [
-      "C:\\Users\\loik.meylan\\.dotnet\\store\\|arch|\\|tfm|",
-      "C:\\Users\\loik.meylan\\.nuget\\packages"
+      "C:\\Users\\benja\\.dotnet\\store\\|arch|\\|tfm|",
+      "C:\\Users\\benja\\.nuget\\packages",
+      "C:\\Program Files\\dotnet\\sdk\\NuGetFallbackFolder"
     ]
   }
 }
